Hoist offline-compile device config table into jit_build
Move the supported-products table out of precompile_fw.cpp and into
jit_device_config.{hpp,cpp} as enumerate_offline_compile_device_configs,
so that both firmware precompile and (upcoming) offline kernel compile
share one source of truth for which (arch, core_descriptor,
soc_descriptor) tuples are officially supported. The descriptor struct
and the YAML table itself stay file-private; consumers see a single
for_each helper that takes a RunTimeOptions reference. The naming
intentionally drops "JIT" from the new symbols because both consumers
want offline compilation as the outcome.

precompile_fw::main() collapses to one call into the new helper.
Verified behavior-preserving: re-running the precompile_fw binary
reports 488/488 JIT cache hits and produces an identical 156-entry
tt_metal/pre-compiled/<build_key>/ directory set before and after.
